requirements.txt
# Core dependencies
langchain
python-dotenv
langchain-google-genai
fastapi[standard]
langchain-community
google-generativeai
Pillow
python-multipart
pydantic
pydantic-settings
uvicorn

# # Testing dependencies
# pytest==8
# pytest-asyncio==0.23.5
# httpx==0.27.0

# # Development dependencies
# black==24.2.0
# flake8==7.0.0